completed all spec functionalities

+++ src/main/java/uk/ac/ed/inf/DronePath/FlightPath.java
@@ -1,5 +1,5 @@
 package uk.ac.ed.inf.DronePath;
 
 public record FlightPath (String orderNo, double fromLongitude, double fromLatitude, double angle,
-                          double toLongitude, double toLatitude, int ticksSinceStartOfCalculation) {
+                          double toLongitude, double toLatitude, long ticksSinceStartOfCalculation) {
 }
